Изменения к дз module1-task3
