+++ src/main.js
@@ -1,16 +1,14 @@
-import NewTaskTripInfoView from './view/trip-info.js';
-import NewTaskFormView from './view/trip-filters.js';
-import NewTaskFilterButtonView from './view/filter-button.js';
-import BoardPresenter from './view/presenter.js';
-import {render} from './render.js';
+import TripInfoView from './view/trip-info-view.js';
+import TripFiltersView from './view/trip-filters-view.js';
+import BoardPresenter from './view/presenter-view.js';
+import {render, RenderPosition} from './render.js';
 
-const siteMainElement = document.querySelector('.trip-main');
-const siteFormElement = siteMainElement.querySelector('.trip-controls__filters');
-const siteSortingElement = document.querySelector('.trip-events');
-const boardPresenter = new BoardPresenter({boardContainer: siteSortingElement});
+const siteHeaderElement = document.querySelector('.trip-main');
+const siteFiltersElement = document.querySelector('.trip-controls__filters');
+const boardElement = document.querySelector('.trip-events');
+const boardPresenter = new BoardPresenter({boardContainer: boardElement});
 
-render(new NewTaskTripInfoView(), siteMainElement);
-render(new NewTaskFormView(), siteFormElement);
-render(new NewTaskFilterButtonView(), siteFormElement);
+render(new TripFiltersView(), siteFiltersElement);
+render(new TripInfoView(), siteHeaderElement, RenderPosition.AFTERBEGIN);
 
 boardPresenter.init();
